+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - `state` do not add dot to empty prefix at own metrics.
 
+### Removed
+- **-TLS** argument in favor of **-tlsOutput** and **-tlsInput**.
+
 ### Added
 - Add compression for both `receiver` and `sender`.
+- New arguments: **-tlsOutput**, **-tlsInput**, **-tlsInputCert**, **-tlsInputKey**.
 - `state` new metrics: **OutBytes**, **OutBpm**.
 
 ## [1.3.0] - 2021-06-03